Update schematic to KiCad 7
This also sets the edge clearance to 0,5mm. This is a new setting, that
was previously set using edge cuts line widths and/or zone clearance
settings. This commit also refills the zones, which are slightly bigger
than before.

This also removes the old .sch and .pro file, which are replaced by
.kicad_sch and .kicad_pro, and removes the lib cache, which is no longer
used in the new schematic format.