Update to KiCad 8
This is the result of opening the project, schematic and board and the
schematic setup and board setup dialogs (and clicking open each section)
without making any changes. In addition, a zone fill is done.

This commit changes all files completely, since the new KiCad adds more
linebreaks to the files it writes.